+++ py/desispec/scripts/zcatalog.py
@@ -198,7 +198,13 @@ def read_redrock(rrfile, group=None, recoadd_fibermap=False, minimal=False, pert
                 index=icol, name='NIGHT')
     elif group == 'cumulative':
         if 'LASTNIGHT' not in data.colnames:
-            data.add_column(np.full(nrows, hdr['NIGHT'], dtype=np.int32),
+            try:
+                lastnight = int(hdr['NIGHT'])
+            except KeyError:
+                # Some daily reductions do not have this set, use the filename.
+                log.warning(f'NIGHT keyword missing from {rrfile}!')
+                lastnight = int(rrfile.split('-')[-1].split('.')[0].replace('thru', ''))
+            data.add_column(np.full(nrows, lastnight, dtype=np.int32),
                     index=icol, name='LASTNIGHT')
     elif group == 'healpix':
         data.add_column(np.full(nrows, hdr['HPXPIXEL'], dtype=np.int32),
@@ -217,11 +223,16 @@ def read_redrock(rrfile, group=None, recoadd_fibermap=False, minimal=False, pert
                 dtype = np.int64
         else:
             dtype = None
-
-        data.add_column(np.full(nrows, hdr['SPGRPVAL'], dtype=dtype),
-                index=icol, name='SPGRPVAL')
     else:
         log.warning(f'SPGRPVAL keyword missing from {rrfile}')
+        if group == 'cumulative':
+            val = lastnight
+            dtype = np.int32
+        else:
+            # This is temporary. The whole section above could do with some refactoring.
+            raise NotImplementedError(f'No method to reconstruct SPGRPVAL!')
+    data.add_column(np.full(nrows, val, dtype=dtype),
+            index=icol, name='SPGRPVAL')
 
     return data, expfibermap